style(fix): fix style UX
