+++ packages/style/scss/components/sub-navigation.scss
@@ -23,13 +23,8 @@
                     padding: 0 40px;
 
                     &-label {
-                        @extend .body-m;
-                        padding: 24px 0 8px;
-
-                        &-solo {
-                            @extend .body-m;
-                            padding: 24px 0;
-                        }
+                        @extend .body-l;
+                        padding: 16px 0;
                     }
                 }
                 &.disabled {
@@ -41,7 +36,6 @@
             &-description {
                 display: block;
                 padding: 0 40px 24px;
-                line-height: $sub-navigation-line-height;
             }
 
             &.mod-selected {
